requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.1
supabase==2.7.4
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.27.2
asyncio==3.4.3
aiofiles==23.2.1
python-multipart==0.0.20
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
websockets>=12.0
# Enhanced system dependencies
cachetools==5.3.0
structlog==23.2.0
aiocache==0.12.0
